backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.1
python-dotenv==1.0.0
moviepy==1.0.3
numpy==1.26.4
pydub==0.25.1
python-magic==0.4.27
werkzeug==3.0.1
gunicorn==21.2.0
pytesseract==0.3.10
opencv-python==4.8.0.74
tensorflow==2.19.0
transformers==4.36.2
torch==2.7.1
openai-whisper==20240930
librosa==0.10.1
scipy==1.11.4
numba==0.58.1
SpeechRecognition==3.10.0

